qr_secret_messages/requirements.txt
pycryptodome==3.19.0
qrcode[pil]==7.4.2
pyzbar==0.1.9
Pillow==10.1.0

